tw/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
llama-cpp-python==0.2.20
duckduckgo_search==3.9.3

# Optional but recommended packages for better performance
transformers==4.35.2
sentence-transformers==2.2.2
torch==2.1.1
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0 